Rename controller implementation files.
